refactor: move project to React Router V7 pre-release
