+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+	"editor.defaultFormatter": "biomejs.biome",
+	"typescript.tsdk": "node_modules/typescript/lib",
+	"[json]": {
+		"editor.defaultFormatter": "biomejs.biome"
+	}
+}
