+++ app/routes.ts
@@ -1,0 +1,4 @@
+import type { RouteConfig } from "@react-router/dev/routes";
+import { index } from "@react-router/dev/routes";
+
+export const routes: RouteConfig = [index("routes/index.tsx")];
