+++ package.json
@@ -1,38 +1,36 @@
 {
-  "name": "iat-test",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "iat-test",
+	"private": true,
+	"sideEffects": false,
+	"type": "module",
+	"scripts": {
+		"dev": "react-router dev",
+		"build": "react-router build",
+		"format": "biome format --write",
+		"start": "react-router-serve ./build/server/index.js",
+		"typecheck": "react-router typegen && tsc"
+	},
+	"dependencies": {
+		"@react-router/node": "7.0.0-pre.0",
+		"@react-router/serve": "7.0.0-pre.0",
+		"isbot": "^5.1.17",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-router": "7.0.0-pre.0"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "^1.9.3",
+		"@react-router/dev": "7.0.0-pre.0",
+		"@types/react": "^18.3.9",
+		"@types/react-dom": "^18.3.0",
+		"autoprefixer": "^10.4.20",
+		"postcss": "^8.4.47",
+		"tailwindcss": "^3.4.13",
+		"typescript": "^5.6.2",
+		"vite": "^5.4.8",
+		"vite-tsconfig-paths": "^5.0.1"
+	},
+	"engines": {
+		"node": ">=20.0.0"
+	}
 }
