--- src/App.css
@@ -1,104 +1,0 @@
-/* App.css */
-
-.IAT-container {
-    width: 80%;
-    margin: 0 auto;
-    border: 2px solid #ddd;
-    border-radius: 10px;
-    margin-top: 100px;
-    padding: 20px;
-    text-align: center;
-    font-family: Arial, sans-serif;
-  }
-  
-  /* Top row instructions */
-  .instructions-row {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 20px;
-    font-size: 1.2rem;
-    color: black;
-    padding: 0 10px;
-  }
-  
-  .instructions-row span {
-    width: 50%;
-    text-align: center;
-  }
-  
-  /* Categories and stimulus row */
-  .categories-stimulus-row {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: 20px 0;
-    padding: 0 10px;
-  }
-  
-  /* Left and Right Category sections */
-  .category-left, .category-right {
-    width: 30%;
-    text-align: left;
-  }
-  
-  .category-left h2, .category-right h2 {
-    font-size: 2rem;
-    color: blue;
-    margin-bottom: 0.5rem;
-  }
-  
-  .category-left h3, .category-right h3 {
-    font-size: 1.5rem;
-    color: green;
-  }
-  
-  /* Central stimulus */
-  .stimulus {
-    width: 40%;
-    text-align: center;
-  }
-  
-  .stimulus h1 {
-    font-size: 2.5rem;
-    color: black;
-  }
-  
-  /* Bottom error instructions */
-  .error-instructions {
-    margin-top: 20px;
-    font-size: 1rem;
-    color: black;
-  }
-  
-  .error-text {
-    color: red;
-    font-weight: bold;
-  }
-  
-  .IAT-container h2 {
-    font-size: 2rem;
-    margin-bottom: 20px;
-  }
-  
-  .IAT-container p {
-    font-size: 1.2rem;
-    margin: 10px 0;
-  }
-
-  .timer-bar {
-    width: 100%; 
-    height: 20px; 
-    background-color: #76c7c0; 
-    border-radius: 10px; 
-    margin-top: 50px;
-    margin-bottom: 20px; 
-  }
-  
-  .timer-fill {
-    height: 100%; 
-    background-color: #ff7694; 
-    border-radius: 10px; 
-    transition: width 0.1s ease;
-  }
-
-
